requirements.txt
Flask>=3.1,<4
Flask-SQLAlchemy>=3.1,<4
gunicorn>=25,<26
openpyxl>=3.1,<4
